Extract shared Discord REST helpers
