+++ src/discord/rest/webhooks/editMessage.ts
@@ -1,0 +1,45 @@
+import logger from "@logger";
+import type { RESTPatchAPIWebhookWithTokenMessageJSONBody } from "discord-api-types/v10";
+
+interface EditDiscordMessageArgs {
+  applicationId: string;
+  token: string;
+  messageId: string;
+  payload: RESTPatchAPIWebhookWithTokenMessageJSONBody;
+}
+
+export const editMessage = async ({
+  applicationId,
+  token,
+  messageId = "@original",
+  payload,
+}: EditDiscordMessageArgs): Promise<void> => {
+  const url = `https://discord.com/api/v10/webhooks/${applicationId}/${token}/messages/${messageId}`;
+
+  logger.debug(
+    { url: url.replace(token, "[TOKEN]"), messageId },
+    "Editing Discord message",
+  );
+
+  const response = await fetch(url, {
+    method: "PATCH",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(payload),
+  });
+
+  if (!response.ok) {
+    const errorText = await response.text();
+    logger.error(
+      { status: response.status, error: errorText, messageId },
+      "Discord message PATCH failed",
+    );
+    throw new Error(
+      `Discord API error: ${response.status} ${response.statusText}`,
+    );
+  }
+
+  logger.debug(
+    { status: response.status, messageId },
+    "Successfully edited Discord message",
+  );
+};
